fix: return the same value when resolving request multiple times
Before this commit, when calling fetchr on the client using the
promise API and resolving it multiple times would lead to undefined
values.

We now make sure to always return the fetch promise in
httpRequest. But doing so, broke the abort behavior since we would not
return an FetchrHttpRequest to the user from fetchr methods. To fix
that, we now save an instance of FetchrHttpRequest in the Request
class and added an abort method to the later one to call
FetchrHttpRequest.abort method.